pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gdrive-explorer"
version = "0.1.0"
description = "A Python CLI application to explore Google Drive and analyze storage usage"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "Topic :: System :: Filesystems",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "google-api-python-client==2.144.0",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.1.0",
    "click==8.1.7",
    "rich==13.7.1",
    "pydantic==2.8.2",
    "pandas==2.2.2",
    "pyyaml==6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.2",
    "black==24.8.0",
    "mypy==1.11.1",
    "pre-commit==3.8.0",
]

[project.scripts]
gdrive-explorer = "gdrive_explorer.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[[tool.mypy.overrides]]
module = [
    "googleapiclient.*",
    "google_auth_oauthlib.*",
    "google.auth.*"
]
ignore_missing_imports = true